atom: refactor AtomMeta to delay the class creation till all work is complete
Doing so will allow using __init_subclass__ on atom subclass. Previously one would get a class missing in particular __atom_members__ making it hard to do anything meaningful (such as customizing pickling).
